+++ libraries/transformer/src/main/java/androidx/media3/transformer/InAppMp4Muxer.java
@@ -122,6 +122,11 @@ public final class InAppMp4Muxer implements Muxer {
       }
       return ImmutableList.of();
     }
+
+    @Override
+    public boolean supportsWritingNegativeTimestampsInEditList() {
+      return true;
+    }
   }
 
   public static final String MUXER_NAME =